possibly fix issue #13 by explicit enum cast to int

+++ src/ConsoleGUI.cpp
@@ -54,9 +54,9 @@ void ConsoleGUI::update()
 {
     for (uint32_t i = 0; i < textHeight; i++) 
     {
-        wattrset(winPtr, COLOR_PAIR(Color::WINDOW_FRAME) | A_REVERSE);
+        wattrset(winPtr, COLOR_PAIR(static_cast<int>(Color::WINDOW_FRAME)) | A_REVERSE);
         mvwprintw(winPtr, (int)i, 0, ">");
-        wattrset(winPtr, COLOR_PAIR(Color::DEF_DEF) | A_REVERSE);
+        wattrset(winPtr, COLOR_PAIR(static_cast<int>(Color::DEF_DEF)) | A_REVERSE);
         mvwprintw(winPtr, (int)i, 1, " ");
         string txt;
         if (i < textBuffer.size()) {
